refac: clean code
