+++ internal/internal.go
@@ -1,9 +1,7 @@
 package internal
 
 import (
-	"bytes"
 	"context"
-	"database/sql"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -11,39 +9,17 @@ import (
 	"runtime"
 	"strconv"
 	"strings"
-	"sync"
 	"time"
 
-	"github.com/getsentry/sentry-go"
-	"github.com/gin-gonic/gin"
 	"github.com/go-co-op/gocron"
 	"github.com/hashicorp/go-version"
 	log "github.com/sirupsen/logrus"
-	"github.com/slack-go/slack"
 	"github.com/uptrace/bun/dialect"
-	"github.com/uptrace/bun/dialect/pgdialect"
-	"github.com/uptrace/bun/dialect/sqlitedialect"
-	"github.com/uptrace/bun/driver/pgdriver"
-	"github.com/uptrace/bun/driver/sqliteshim"
 
 	"github.com/spf13/cobra"
-	v1 "github.com/tailwarden/komiser/internal/api/v1"
 	"github.com/tailwarden/komiser/internal/config"
 	"github.com/tailwarden/komiser/models"
-	"github.com/tailwarden/komiser/providers"
-	"github.com/tailwarden/komiser/providers/aws"
-	"github.com/tailwarden/komiser/providers/azure"
-	"github.com/tailwarden/komiser/providers/civo"
-	do "github.com/tailwarden/komiser/providers/digitalocean"
-	"github.com/tailwarden/komiser/providers/gcp"
-	"github.com/tailwarden/komiser/providers/k8s"
-	"github.com/tailwarden/komiser/providers/linode"
-	"github.com/tailwarden/komiser/providers/mongodbatlas"
-	"github.com/tailwarden/komiser/providers/oci"
-	"github.com/tailwarden/komiser/providers/ovh"
 
-	"github.com/tailwarden/komiser/providers/scaleway"
-	"github.com/tailwarden/komiser/providers/tencent"
 	"github.com/tailwarden/komiser/utils"
 	"github.com/uptrace/bun"
 )
@@ -62,12 +38,12 @@ func Exec(address string, port int, configPath string, telemetry bool, a utils.A
 
 	ctx := context.Background()
 
-	cfg, clients, accounts, err := config.Load(configPath, telemetry, analytics, db)
+	cfg, clients, accounts, err := config.Load(configPath, telemetry, analytics)
 	if err != nil {
 		return err
 	}
 
-	err = setupDBConnection(cfg)
+	err = utils.SetupDBConnection(cfg, db)
 	if err != nil {
 		return err
 	}
@@ -91,16 +67,17 @@ func Exec(address string, port int, configPath string, telemetry bool, a utils.A
 		}
 
 		_, err = cron.Every(1).Hours().Do(func() {
-			alertsExist, alerts := checkIfAlertsExist(ctx)
-
-			if alertsExist {
+			alerts, err := listAlerts(ctx)
+			if err != nil {
+				log.WithError(err).Error("failed to list alerts")
+			}
+			if len(alerts) > 0 {
 				log.Info("Checking Alerts")
 				checkingAlerts(ctx, *cfg, telemetry, port, alerts)
 			}
 		})
-
 		if err != nil {
-			log.WithError(err).Error("setting up cron job failed")
+			log.WithError(err).Error("failed to setup alert checking cron job")
 		}
 
 		_, err = cron.Every(1).Friday().At("09:00").Do(func() {
@@ -110,9 +87,8 @@ func Exec(address string, port int, configPath string, telemetry bool, a utils.A
 				sendCostBreakdownReport(ctx, *cfg)
 			}
 		})
-
 		if err != nil {
-			log.WithError(err).Error("setting up cron job failed")
+			log.WithError(err).Error("failed to setup cron job")
 		}
 
 		cron.StartAsync()
@@ -124,187 +100,7 @@ func Exec(address string, port int, configPath string, telemetry bool, a utils.A
 	if err != nil {
 		return err
 	}
-
 	return nil
-}
-
-func checkIfAlertsExist(ctx context.Context) (bool, []models.Alert) {
-	alerts := make([]models.Alert, 0)
-
-	err := db.NewRaw("SELECT * FROM alerts").Scan(ctx, &alerts)
-	if err != nil {
-		log.WithError(err).Error("scan failed")
-	}
-	if len(alerts) > 0 {
-		return true, alerts
-	}
-
-	return false, alerts
-}
-
-func loggingMiddleware() gin.HandlerFunc {
-	return func(ctx *gin.Context) {
-		startTime := time.Now()
-		ctx.Next()
-		endTime := time.Now()
-		latencyTime := endTime.Sub(startTime)
-		reqMethod := ctx.Request.Method
-		reqUri := ctx.Request.RequestURI
-		statusCode := ctx.Writer.Status()
-		clientIP := ctx.ClientIP()
-
-		log.WithFields(log.Fields{
-			"method":  reqMethod,
-			"uri":     reqUri,
-			"status":  statusCode,
-			"latency": latencyTime,
-			"ip":      clientIP,
-		}).Info("HTTP request")
-
-		ctx.Next()
-	}
-}
-
-func runServer(address string, port int, telemetry bool, cfg models.Config, configPath string, accounts []models.Account) error {
-	log.Infof("Komiser version: %s, commit: %s, buildt: %s", Version, Commit, Buildtime)
-
-	r := v1.Endpoints(context.Background(), telemetry, analytics, db, cfg, configPath, accounts)
-
-	r.Use(loggingMiddleware())
-
-	if err := r.Run(fmt.Sprintf("%s:%d", address, port)); err != nil {
-		return err
-	}
-
-	log.Infof("Server started on %s:%d", address, port)
-
-	return nil
-}
-
-func setupDBConnection(c *models.Config) error {
-	var sqldb *sql.DB
-	var err error
-
-	if len(c.SQLite.File) == 0 && len(c.Postgres.URI) == 0 {
-		log.Println("Database wasn't configured yet")
-		return nil
-	}
-
-	if len(c.SQLite.File) > 0 {
-		sqldb, err = sql.Open(sqliteshim.ShimName, fmt.Sprintf("file:%s?cache=shared", c.SQLite.File))
-		if err != nil {
-			return err
-		}
-		sqldb.SetMaxIdleConns(1000)
-		sqldb.SetConnMaxLifetime(0)
-
-		db = bun.NewDB(sqldb, sqlitedialect.New())
-
-		log.Println("Data will be stored in SQLite")
-	} else {
-		sqldb = sql.OpenDB(pgdriver.NewConnector(pgdriver.WithDSN(c.Postgres.URI)))
-		db = bun.NewDB(sqldb, pgdialect.New())
-		log.Println("Data will be stored in PostgreSQL")
-	}
-
-	return nil
-}
-
-func triggerFetchingWorkflow(ctx context.Context, client providers.ProviderClient, provider string, telemetry bool, regions []string, wp *providers.WorkerPool) {
-	localHub := sentry.CurrentHub().Clone()
-
-	defer func() {
-		err := recover()
-		if err != nil {
-			log.WithField("err", err).Error(fmt.Sprintf("error fetching %s resources", provider))
-			localHub.CaptureException(err.(error))
-			localHub.Flush(2 * time.Second)
-		}
-	}()
-
-	localHub.ConfigureScope(func(scope *sentry.Scope) {
-		scope.SetTag("provider", provider)
-	})
-
-	if telemetry {
-		analytics.TrackEvent("fetching_resources", map[string]interface{}{
-			"provider": provider,
-		})
-	}
-
-	switch provider {
-	case "AWS":
-		aws.FetchResources(ctx, client, regions, db, telemetry, analytics, wp)
-	case "DigitalOcean":
-		do.FetchResources(ctx, client, db, telemetry, analytics, wp)
-	case "OCI":
-		oci.FetchResources(ctx, client, db, telemetry, analytics, wp)
-	case "Civo":
-		civo.FetchResources(ctx, client, db, telemetry, analytics, wp)
-	case "Kubernetes":
-		k8s.FetchResources(ctx, client, db, telemetry, analytics, wp)
-	case "Linode":
-		linode.FetchResources(ctx, client, db, telemetry, analytics, wp)
-	case "Tencent":
-		tencent.FetchResources(ctx, client, db, telemetry, analytics, wp)
-	case "Azure":
-		azure.FetchResources(ctx, client, db, telemetry, analytics, wp)
-	case "Scaleway":
-		scaleway.FetchResources(ctx, client, db, telemetry, analytics, wp)
-	case "MongoDBAtlas":
-		mongodbatlas.FetchResources(ctx, client, db, telemetry, analytics, wp)
-	case "GCP":
-		gcp.FetchResources(ctx, client, db, telemetry, analytics, wp)
-	case "OVH":
-		ovh.FetchResources(ctx, client, db, telemetry, analytics, wp)
-	}
-}
-
-func fetchResources(ctx context.Context, clients []providers.ProviderClient, regions []string, telemetry bool) {
-	numWorkers := 64
-	wp := providers.NewWorkerPool(numWorkers)
-	wp.Start()
-
-	var wwg sync.WaitGroup
-	workflowTrigger := func(client providers.ProviderClient, provider string) {
-		wwg.Add(1)
-		go func() {
-			defer wwg.Done()
-			triggerFetchingWorkflow(ctx, client, provider, telemetry, regions, wp)
-		}()
-	}
-
-	for _, client := range clients {
-		if client.AWSClient != nil {
-			workflowTrigger(client, "AWS")
-		} else if client.DigitalOceanClient != nil {
-			workflowTrigger(client, "DigitalOcean")
-		} else if client.OciClient != nil {
-			workflowTrigger(client, "OCI")
-		} else if client.CivoClient != nil {
-			workflowTrigger(client, "Civo")
-		} else if client.K8sClient != nil {
-			workflowTrigger(client, "Kubernetes")
-		} else if client.LinodeClient != nil {
-			workflowTrigger(client, "Linode")
-		} else if client.TencentClient != nil {
-			workflowTrigger(client, "Tencent")
-		} else if client.AzureClient != nil {
-			workflowTrigger(client, "Azure")
-		} else if client.ScalewayClient != nil {
-			workflowTrigger(client, "Scaleway")
-		} else if client.MongoDBAtlasClient != nil {
-			workflowTrigger(client, "MongoDBAtlas")
-		} else if client.GCPClient != nil {
-			workflowTrigger(client, "GCP")
-		} else if client.OVHClient != nil {
-			workflowTrigger(client, "OVH")
-		}
-		log.Println("Workflow triggered for client:", client.Name)
-	}
-
-	wwg.Wait()
-	wp.Wait()
 }
 
 func checkUpgrade() {
@@ -339,152 +135,6 @@ func checkUpgrade() {
 			if v1.LessThan(v2) {
 				log.Warnf("Newer Komiser version is available: %s", target.Version)
 				log.Warnf("Upgrade instructions: https://github.com/tailwarden/komiser")
-			}
-		}
-	}
-}
-
-func hitCustomWebhook(endpoint string, secret string, viewName string, resources int, cost float64, alertType string) {
-	var payloadJSON []byte
-	var err error
-	payload := models.CustomWebhookPayload{
-		Komiser:   Version,
-		View:      viewName,
-		Timestamp: time.Now().Unix(),
-	}
-
-	switch alertType {
-	case "BUDGET":
-		payload.Message = "Cost alert"
-		payload.Data = cost
-	case "USAGE":
-		payload.Message = "Usage alert"
-		payload.Data = float64(resources)
-	default:
-		log.Error("Invalid Alert Type")
-		return
-	}
-
-	payloadJSON, err = json.Marshal(payload)
-	if err != nil {
-		log.Error("Couldn't encode JSON payload:", err)
-		return
-	}
-
-	req, err := http.NewRequest("POST", endpoint, bytes.NewBuffer(payloadJSON))
-	if err != nil {
-		log.Error("Couldn't create HTTP request for custom webhook endpoint:", err)
-		return
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-	if len(secret) > 0 {
-		req.Header.Set("Authorization", secret)
-	}
-
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		log.Error("Couldn't make HTTP request for custom webhook endpoint:", err)
-		return
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != 200 {
-		log.Error("Custom Webhook with endpoint " + endpoint + " returned back a status code of " + string(rune(resp.StatusCode)) + " . Expected Status Code: 200")
-		return
-	}
-}
-
-func hitSlackWebhook(viewName string, port int, viewId int, resources int, cost float64, webhookUrl string, alertType string) {
-	attachment := slack.Attachment{
-		Color:         "danger",
-		AuthorName:    "Komiser",
-		AuthorSubname: "by Tailwarden",
-		AuthorLink:    "https://tailwarden.com",
-		AuthorIcon:    "https://cdn.komiser.io/images/komiser-logo.jpeg",
-		Footer:        "Komiser",
-		Actions: []slack.AttachmentAction{
-			{
-				Name: "open",
-				Text: "Open view",
-				Type: "button",
-				URL:  fmt.Sprintf("http://localhost:%d/inventory?view=%d", port, viewId),
-			},
-		},
-		Fields: []slack.AttachmentField{
-			{
-				Title: "View",
-				Value: viewName,
-			},
-		},
-		FooterIcon: "https://github.com/tailwarden/komiser",
-		Ts:         json.Number(strconv.FormatInt(time.Now().Unix(), 10)),
-	}
-
-	if alertType == "BUDGET" {
-		attachment.Text = "Cost alert :warning:"
-		attachment.Fields = append(attachment.Fields, slack.AttachmentField{
-			Title: "Cost",
-			Value: fmt.Sprintf("%.2f$", cost),
-		})
-	} else if alertType == "USAGE" {
-		attachment.Text = "Usage alert :warning:"
-		attachment.Fields = append(attachment.Fields, slack.AttachmentField{
-			Title: "Resources",
-			Value: fmt.Sprintf("%d", resources),
-		})
-	}
-
-	msg := slack.WebhookMessage{
-		Attachments: []slack.Attachment{attachment},
-	}
-
-	err := slack.PostWebhook(webhookUrl, &msg)
-	if err != nil {
-		log.Warn(err)
-	}
-
-}
-
-func checkingAlerts(ctx context.Context, cfg models.Config, telemetry bool, port int, alerts []models.Alert) {
-	for _, alert := range alerts {
-		var view models.View
-		err := db.NewRaw(fmt.Sprintf("SELECT * FROM views WHERE id = %s", alert.ViewId)).Scan(ctx, &view)
-		if err != nil {
-			log.WithError(err).Error("scan failed")
-		}
-
-		stats, err := getViewStats(ctx, view.Filters)
-		if err != nil {
-			log.Error("Couldn't get stats for view:", view.Name)
-		} else {
-			if alert.Type == "BUDGET" && alert.Budget <= stats.Costs {
-				if telemetry {
-					analytics.TrackEvent("sending_alerts", map[string]interface{}{
-						"type": "budget",
-					})
-				}
-				if alert.IsSlack {
-					log.Info("Sending Slack budget alert for view:", view.Name)
-					hitSlackWebhook(view.Name, port, int(view.Id), 0, stats.Costs, cfg.Slack.Webhook, alert.Type)
-				} else {
-					log.Info("Sending Custom Webhook budget alert for view:", view.Name)
-					hitCustomWebhook(alert.Endpoint, alert.Secret, view.Name, 0, stats.Costs, alert.Type)
-				}
-			}
-			if alert.Type == "USAGE" && alert.Usage <= stats.Resources {
-				if telemetry {
-					analytics.TrackEvent("sending_alerts", map[string]interface{}{
-						"type": "usage",
-					})
-				}
-				if alert.IsSlack {
-					log.Info("Sending Slack usage alert for view:", view.Name)
-					hitSlackWebhook(view.Name, port, int(view.Id), stats.Resources, 0, cfg.Slack.Webhook, alert.Type)
-				} else {
-					log.Info("Sending Custom Webhook usage alert for view:", view.Name)
-					hitCustomWebhook(alert.Endpoint, alert.Secret, view.Name, stats.Resources, 0, alert.Type)
-				}
 			}
 		}
 	}
@@ -566,7 +216,7 @@ func getViewStats(ctx context.Context, filters []models.Filter) (models.ViewStat
 					whereQueries = append(whereQueries, fmt.Sprintf("((res->>'key' = '%s') AND (res->>'value' != ''))", key))
 				}
 			default:
-				return models.ViewStat{}, errors.New("Operation is invalid or not supported")
+				return models.ViewStat{}, errors.New("operation is invalid or not supported")
 			}
 		} else if filter.Field == "tags" {
 			switch filter.Operator {
@@ -583,7 +233,7 @@ func getViewStats(ctx context.Context, filters []models.Filter) (models.ViewStat
 					whereQueries = append(whereQueries, "jsonb_array_length(tags) != 0")
 				}
 			default:
-				return models.ViewStat{}, errors.New("Operation is invalid or not supported")
+				return models.ViewStat{}, errors.New("operation is invalid or not supported")
 			}
 		} else if filter.Field == "cost" {
 			switch filter.Operator {
@@ -683,118 +333,5 @@ func getViewStats(ctx context.Context, filters []models.Filter) (models.ViewStat
 		}
 
 		return output, nil
-	}
-}
-
-func sendTagsCoverageReport(ctx context.Context, cfg models.Config) {
-	tags := make([]struct {
-		Total int        `bun:"total"`
-		Label models.Tag `bun:"label"`
-	}, 0)
-
-	err := db.NewRaw("SELECT count(*) as total, value as label FROM resources CROSS JOIN json_each(tags) GROUP BY value ORDER BY total DESC").Scan(ctx, &tags)
-	if err != nil {
-		log.WithError(err).Error("scan failed")
-	}
-
-	fields := make([]slack.AttachmentField, 0)
-
-	for _, tag := range tags {
-		fields = append(fields, slack.AttachmentField{
-			Title: fmt.Sprintf("%s:%s", tag.Label.Key, tag.Label.Value),
-			Value: fmt.Sprintf("%d", tag.Total),
-			Short: true,
-		})
-	}
-
-	output := struct {
-		Total int `bun:"total"`
-	}{}
-
-	err = db.NewRaw("SELECT COUNT(*) as total FROM resources where json_array_length(tags) = 0;").Scan(ctx, &output)
-	if err != nil {
-		log.WithError(err).Error("scan failed")
-	}
-
-	currentTime := time.Now()
-
-	attachment := slack.Attachment{
-		Color:         "good",
-		AuthorName:    "Komiser",
-		AuthorSubname: "by Tailwarden",
-		AuthorLink:    "https://tailwarden.com",
-		AuthorIcon:    "https://cdn.komiser.io/images/komiser-logo.jpeg",
-		Text:          fmt.Sprintf("On %s %d: *%d* of your resources are untagged. Below list of most used key/value pairs:", currentTime.Month(), currentTime.Day(), output.Total),
-		Footer:        "Komiser",
-		Fields:        fields,
-		FooterIcon:    "https://github.com/tailwarden/komiser",
-		Ts:            json.Number(strconv.FormatInt(time.Now().Unix(), 10)),
-	}
-	msg := slack.WebhookMessage{
-		Attachments: []slack.Attachment{attachment},
-	}
-
-	err = slack.PostWebhook(cfg.Slack.Webhook, &msg)
-	if err != nil {
-		log.Warn(err)
-	}
-}
-
-func sendCostBreakdownReport(ctx context.Context, cfg models.Config) {
-	groups := make([]models.OutputCostByField, 0)
-	currentTime := time.Now()
-
-	for _, field := range []string{"service", "provider", "account", "region"} {
-		err := db.NewRaw(fmt.Sprintf("SELECT %s as label, SUM(cost) as total FROM resources GROUP BY %s ORDER by total desc;", field, field)).Scan(ctx, &groups)
-		if err != nil {
-			log.WithError(err).Error("scan failed")
-		}
-
-		segments := groups
-
-		if len(groups) > 3 {
-			segments = groups[:4]
-			if len(groups) > 4 {
-				sum := 0.0
-				for i := 4; i < len(groups); i++ {
-					sum += groups[i].Total
-				}
-
-				segments = append(segments, models.OutputCostByField{
-					Label: "Others",
-					Total: sum,
-				})
-			}
-		}
-
-		fields := make([]slack.AttachmentField, 0)
-		for _, segment := range segments {
-			fields = append(fields, slack.AttachmentField{
-				Title: segment.Label,
-				Value: fmt.Sprintf("%.2f", segment.Total),
-				Short: true,
-			})
-		}
-
-		attachment := slack.Attachment{
-			Color:         "good",
-			AuthorName:    "Komiser",
-			AuthorSubname: "by Tailwarden",
-			AuthorLink:    "https://tailwarden.com",
-			AuthorIcon:    "https://cdn.komiser.io/images/komiser-logo.jpeg",
-			Text:          fmt.Sprintf("On %s %d: cost breakdown by cloud %s", currentTime.Month(), currentTime.Day(), field),
-			Footer:        "Komiser",
-			Fields:        fields,
-			FooterIcon:    "https://github.com/tailwarden/komiser",
-			Ts:            json.Number(strconv.FormatInt(time.Now().Unix(), 10)),
-		}
-		msg := slack.WebhookMessage{
-			Attachments: []slack.Attachment{attachment},
-		}
-
-		err = slack.PostWebhook(cfg.Slack.Webhook, &msg)
-		if err != nil {
-			log.Warn(err)
-		}
 	}
 }
